refactor: expose mcp server functionalities as package
